Small cleanups + got rid of an unsafe idiom
Multiple parts of the code where both advancing a pointer and decrementing
the number of available bytes in lockstep with each other.
Which has an obvious tendency to break if that lockstep isn't maintained
for whatever reason. Comparing the pointer with a pointer to the end of
the memory area is nicer.